Imported data from BBC file in an intermediate form with read.table
I managed to import the file from read.table in  an intermediate form that still needs to be processed to remove the tags "n:" to specify the column. This first require to control wether if each column has different "n:" as string start. If not, that is good. If yes, that is bad as I need a way to properly reallocate.